sharing [nfc]: Move helper kotlin code to a new file.
New file created is called `sharing.SharingHelper.kt`, it contains
all helper methods related to handling SEND intent.

Signature of `handleSend` and `maybeHandleIntent` were required to
be changed during migration. They contain an extra parameter
`application: ReactApplication`.

MainActivity is modified accordingly to pass this parameter.
